refactor(backend): Se reubicaron muchas carpetas y se mejoró la documentación
