+++ backend/src/app/admin/users/repositories/PrismaUsersRepository.ts
@@ -1,0 +1,19 @@
+import { UsersRepository } from "./UsersRepository";
+
+export class PrismaUsersRepository implements UsersRepository {
+    createUser(): void {
+        throw new Error("Method not implemented.");
+    }
+    editUser(): void {
+        throw new Error("Method not implemented.");
+    }
+    deleteUser(): void {
+        throw new Error("Method not implemented.");
+    }
+    getAll(): void {
+        throw new Error("Method not implemented.");
+    }
+    getUser(): void {
+        throw new Error("Method not implemented.");
+    }
+}
