+++ backend/src/app/admin/users/repositories/UsersRepository.ts
@@ -1,0 +1,7 @@
+export interface UsersRepository {
+    createUser(): void
+    editUser(): void
+    deleteUser(): void
+    getAll(): void
+    getUser(): void
+}
