--- backend/src/app/users/repositories/PrismaUsersRepository.ts
@@ -1,19 +1,0 @@
-import { UsersRepository } from "./UsersRepository";
-
-export class PrismaUsersRepository implements UsersRepository {
-    createUser(): void {
-        throw new Error("Method not implemented.");
-    }
-    editUser(): void {
-        throw new Error("Method not implemented.");
-    }
-    deleteUser(): void {
-        throw new Error("Method not implemented.");
-    }
-    getAll(): void {
-        throw new Error("Method not implemented.");
-    }
-    getUser(): void {
-        throw new Error("Method not implemented.");
-    }
-}
