--- backend/src/app/users/repositories/UsersRepository.ts
@@ -1,7 +1,0 @@
-export interface UsersRepository {
-    createUser(): void
-    editUser(): void
-    deleteUser(): void
-    getAll(): void
-    getUser(): void
-}
