--- backend/src/routes/v1/accessRoutes.ts
@@ -1,17 +1,0 @@
-import { Router } from 'express'
-
-export const accessRoutes = () => {
-  const router = Router()
-
-  router.post('/', async (req, res) => {
-    // Registras un Nuevo Acceso
-    res.status(200).send('hola')
-  })
-
-  router.get('/', async (req, res) => {
-    // Obtienes el Historial de Accesos con Paginación
-    res.status(200).send('hola')
-  })
-
-  return router
-}
